requirements.txt
asgiref==3.3.4
autopep8==1.5.6
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
decorator==5.0.9
Django==3.2.4
django-cors-headers==3.7.0
django-storages==1.11.1
httpie==2.4.0
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
mysqlclient==2.0.3
numpy==1.20.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
PyJWT==2.1.0
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
gunicorn==20.1.0
boto3==1.17.98